utils/requirements.txt
psycopg2-binary>=2.9.10
faker>=37.8.0
